small usersettings updates
